resolve return type of sub_0800A868
