+++ include/code_0800A868.h
@@ -4,7 +4,7 @@
 #include "global.h"
 #include "data.h"
 
-extern s32 sub_0800A868(u8 playerId, struct Object *obj);
+extern struct Object2 *sub_0800A868(u8 playerId, struct Object *obj);
 extern void sub_0800A8BC(void);
 extern void sub_0800A8C8(u8 arg0, u8 arg1, u8 arg2);
 extern void sub_0800A8F8(u8 arg0, u8 arg1, u8 arg2);
