+++ src/multi_08019F28.c
@@ -812,7 +812,6 @@ static void sub_0801A970(struct Multi_08019F28 *r4)
     struct Unk_020382A0_sub stru;
 
     r4->unk9C = 0;
-    // This is an LCG
     stru.unk00 = Rand32();
     stru.unk04 = gUnk_0203ADE0;
     sub_08030C94(1, &stru);
